backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
ipaddress==1.0.23
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
